Implement "including earlier instances in this rota" logic
Conditions of the form

    if A > 0
    then B >= 0

Can be translated into ILP:
http://www.yzuda.org/Useful_Links/optimization/if-then-else-02.html

Actually, all conditions can be translated into ILP because it's NP
complete, just like SAT / SMT.  But this form is good enough.

For example, someone can only be a secondary in week N if they have
shadowed at least twice by the start of week N.  This can be modelled
with the help of an auxilliary variable to track how many times
someone has shadowed like so:

    if times_shadowed[week, person] - 1 > 0
    then rota[week, person, secondary] >= 0

Note the "-1", not "-2".  This is because it's a ">", not a ">=".