Revert "Switch to Cbc"
Cbc is returning incorrect solutions.  The mailing list suggests that
this means the model is bad in some way I don't fully
understand (having too large a range of values, perhaps?), and turning
on the magical flag to fix it slows it down as much as gplk - ie, too
long to wait.

This reverts commit 05f8ff4e8b62203b53ec1e6877e771d9c20c572d.