Do not bake constraints 1.2.3 and 1.6.3 into the ILP model
These drastically slow it down when 2.8 (coming up in the next commit)
is present.  I got bored and killed it after 5 minutes.  Even worse,
the solution is invalid when the time limit on glpk is enforced!

However, with these constraints absent, it is solved in a matter of
seconds (with optimisation disabled).  The other constraints are
enough to ensure the generated rota is still valid.