Default to Cbc
Now tht the magic numbers have been shrunk, Cbc seems to reliably
work.  I left it running overnight, recording the "real" time taken
for the cli to generate the GOV.UK rota, here are some stats:

sum     48595.7
count   813
mean    59.7733
median  29.34
min     9.61
max     1639.50

So the median is under 30s - which is the timeout I previously had set
for glpk.  glpk with a timeout (or mipgap) seems to have difficulty
allocating shadows, so it's good that Cbc runs to completion (and
achieves good shadow allocation) in the same time.

The max time of nearly half an hour is less good.